reactor: remove old tee module
