+++ src/error.rs
@@ -4,6 +4,10 @@ use axum::{
 };
 pub struct AppError(anyhow::Error);
 
+use axum::BoxError;
+use serde::{Deserialize, Serialize};
+use std::borrow::Cow;
+
 impl IntoResponse for AppError {
     fn into_response(self) -> Response {
         (
@@ -21,4 +25,27 @@ where
     fn from(err: E) -> Self {
         Self(err.into())
     }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct JsonResponse {
+    code: u16,
+    result: String,
+}
+pub async fn handle_error(error: BoxError) -> impl IntoResponse {
+    if error.is::<tower::timeout::error::Elapsed>() {
+        return (StatusCode::REQUEST_TIMEOUT, Cow::from("request timed out"));
+    }
+
+    if error.is::<tower::load_shed::error::Overloaded>() {
+        return (
+            StatusCode::SERVICE_UNAVAILABLE,
+            Cow::from("service is overloaded, try again later"),
+        );
+    }
+
+    (
+        StatusCode::INTERNAL_SERVER_ERROR,
+        Cow::from(format!("Unhandled internal error: {error}")),
+    )
 }
