--- src/tee/handler.rs
@@ -1,28 +1,0 @@
-use axum::http::StatusCode;
-use axum::response::IntoResponse;
-use axum::BoxError;
-use serde::{Deserialize, Serialize};
-use std::borrow::Cow;
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct JsonResponse {
-    code: u16,
-    result: String,
-}
-pub async fn handle_error(error: BoxError) -> impl IntoResponse {
-    if error.is::<tower::timeout::error::Elapsed>() {
-        return (StatusCode::REQUEST_TIMEOUT, Cow::from("request timed out"));
-    }
-
-    if error.is::<tower::load_shed::error::Overloaded>() {
-        return (
-            StatusCode::SERVICE_UNAVAILABLE,
-            Cow::from("service is overloaded, try again later"),
-        );
-    }
-
-    (
-        StatusCode::INTERNAL_SERVER_ERROR,
-        Cow::from(format!("Unhandled internal error: {error}")),
-    )
-}
