--- src/tee/mod.rs
@@ -1,2 +1,0 @@
-pub mod handler;
-pub mod model;
